+++ backend/bookmarks/views.py
@@ -10,7 +10,10 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from downloads.tasks import download_bookmark
+
 from .models import Bookmark, Folder, Tag
+from .permissions import IsOwnerOnly
 from .serializers import (
     BookmarkCreateFromTelegramSerializer,
     BookmarkCreateFromWebSerializer,
@@ -21,7 +24,6 @@ from .serializers import (
     FolderSerializer,
     TagListSerializer,
 )
-from .permissions import IsOwnerOnly
 from .utils import parse_url_info
 
 
@@ -236,6 +238,10 @@ def bookmark_create_from_web(request: Request) -> Response:
 
         # Return detailed bookmark info to the user:
         detail_serializer = BookmarkListSerializer(instance=bookmark, many=False)
+
+        # Download link
+        download_bookmark.delay(bookmark_id=bookmark.pk)
+
         return Response(detail_serializer.data, status=status.HTTP_201_CREATED)
 
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)